[#64][FEATURE] 리뷰 세부 옵션 추가

+++ src/components/Review/ReviewBox.jsx
@@ -7,25 +7,24 @@ import { db } from '../../../Firebase';
 import { doc, deleteDoc, collection, getDocs } from 'firebase/firestore';
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router';
+import { unstable_renderSubtreeIntoContainer } from 'react-dom';
+import { async } from '@firebase/util';
 
 export default function ReviewBox() {
   const [reviews, setReviews] = useState([]);
-  //로그인한 유저의 리뷰만 저장
-  // const [userReview, setUserReview] = useState([]);
+  const [loading, setLoading] = useState(true);
 
   const navigate = useNavigate();
 
   //로그인한 유저의 uid 가져오기
   const userId = JSON.parse(localStorage.getItem('user')).user.uid;
 
-  const isEdit = (id, reviewContent, recommendVal) => {
-    console.log('넘겨온 userId: ' + id);
-    console.log('넘겨온 content: ' + reviewContent);
-    console.log('넘겨온 추천 여부: ' + recommendVal);
+  const isEdit = (id, userId, reviewContent, recommendVal) => {
     const content = reviewContent;
     const recommend = recommendVal;
     navigate('/editreview', {
       state: {
+        id: id,
         userId: userId,
         content: content,
         recommendVal: recommend,
@@ -33,15 +32,21 @@ export default function ReviewBox() {
     });
   };
 
+  //[사이드이펙트] DB -> 리뷰 요청/응답
   useEffect(() => {
     // 비동기로 데이터 받을준비
     const getReviews = async () => {
-      //db의 reviews 컬렉션 가져오기
-      const reviewsCollectionRef = collection(db, 'reviews');
-      // getDocs로 컬렉션안에 데이터 가져오기
-      const data = await getDocs(reviewsCollectionRef);
-      // reviews에 data안의 자료 추가. 객체에 id 덮어씌우는거
-      setReviews(data.docs.map((doc) => ({ ...doc.data(), id: doc.id })));
+      try {
+        //db의 reviews 컬렉션 가져오기
+        const reviewsCollectionRef = collection(db, 'reviews');
+        // getDocs로 컬렉션안에 데이터 가져오기
+        const data = await getDocs(reviewsCollectionRef);
+        // reviews에 data안의 자료 추가. 객체에 id 덮어씌우는거
+        setReviews(data.docs.map((doc) => ({ ...doc.data(), id: doc.id })));
+      } finally {
+        //리뷰를 모두 받은 경우, 로딩 상태 false로 변경
+        setLoading(false);
+      }
     };
 
     getReviews();
@@ -57,46 +62,64 @@ export default function ReviewBox() {
     if (window.confirm(`${name}님의 데이터를 삭제하시겠습니까?`)) {
       await deleteDoc(reviewsDoc);
     }
+
+    //리뷰 삭제 시 리뷰 조회 페이지 리렌더링
+    window.location.href = '/mypage/review';
   };
 
-  const showReviews = userReview.map((value) => (
-    <div key={value.id}>
-      <ReviewBoxWrapper>
-        <ReviewBoxHeader>
-          <ParkingLot>파킹 주차장</ParkingLot>
-          <BtnWrapper>
-            <ReviewUpdateButton
-              className="btnUpdate"
-              onClick={() => {
-                isEdit(value.userId, value.content, value.recommend);
-              }}
-            />
-            <ReviewDeleteButton
-              onClick={() => {
-                deleteReview(value.id, value.name);
-              }}
-            />
-          </BtnWrapper>
-        </ReviewBoxHeader>
-        <ReviewWrapper>
-          <ReviewInfo>
-            {value.recommend ? <RecommendTag /> : <NotRecommendTag />}
-            <p className="reviewDate">{new Date(value.date).toLocaleString()} </p>
-          </ReviewInfo>
-          <ReviewContent>{value.content}</ReviewContent>
-        </ReviewWrapper>
-      </ReviewBoxWrapper>
-    </div>
-  ));
+  const showReviews = () => {
+    if (loading) {
+      return <ReviewRoading>작성한 리뷰 로딩 중...</ReviewRoading>;
+    } else {
+      if (reviews.filter((value) => value.userId === userId).length > 0) {
+        return userReview
+          .sort((a, b) => b.date - a.date)
+          .map((value) => (
+            <ReviewBoxWrapper key={value.id}>
+              <ReviewBoxHeader>
+                <ParkingLot>파킹 주차장</ParkingLot>
+                <BtnWrapper>
+                  <ReviewUpdateButton
+                    className="btnUpdate"
+                    onClick={() => {
+                      isEdit(value.id, value.userId, value.content, value.recommend);
+                    }}
+                  />
+                  <ReviewDeleteButton
+                    onClick={() => {
+                      deleteReview(value.id, value.name);
+                    }}
+                  />
+                </BtnWrapper>
+              </ReviewBoxHeader>
+              <ReviewWrapper>
+                <ReviewInfo>
+                  {value.recommend ? <RecommendTag /> : <NotRecommendTag />}
+                  <p className="reviewDate">{new Date(value.date).toLocaleString()} </p>
+                </ReviewInfo>
+                <ReviewContent>{value.content}</ReviewContent>
+              </ReviewWrapper>
+            </ReviewBoxWrapper>
+          ));
+      } else {
+        return <NoReview>작성한 리뷰가 없습니다.</NoReview>;
+      }
+    }
+  };
 
-  return <>{showReviews}</>;
+  return <>{showReviews()}</>;
 }
 
 const ReviewBoxWrapper = styled.li`
+  width: 100%;
+  box-sizing: border-box;
   min-width: 320px;
   height: 221px;
   list-style: none;
   padding: 0;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  border-bottom: 1px solid ${theme.colors.orangeMain};
 `;
 
 const ReviewBoxHeader = styled.div`
@@ -130,6 +153,7 @@ const ReviewWrapper = styled.div`
   padding: 20px 16px;
   letter-spacing: -0.65px;
   font-weight: 400;
+  overflow: scroll;
 `;
 
 const ReviewInfo = styled.div`
@@ -144,4 +168,12 @@ const ReviewInfo = styled.div`
 const ReviewContent = styled.p`
   color: ${theme.colors.black};
   line-height: 24px;
+`;
+
+const NoReview = styled.h2`
+  font-size: ${theme.fontSizes.subTitle1};
+`;
+
+const ReviewRoading = styled.h2`
+  font-size: ${theme.fontSizes.subTitle1};
 `;
